Relocate exporter package
